#8 [2021 UPDATE] USER PROFILE

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video.js"
+import User from "../models/User.js"
 
 export const home = async (req, res) => {
     const videos = await Video.find({}).sort({ createdAt: "desc" });
@@ -7,7 +8,9 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+
+    // populate("owner") Video 모델에 ref: "User" 모델을 참조하겠다고 선언했기 때문에 관계 객체를 형성합니다.
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.status(404).render("error/404", { pageTitle: "Video not found." });    
     }
@@ -15,20 +18,28 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
+    const { user: { _id } } = req.session;
     const { id } = req.params;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("error/404", { pageTitle: "Video not found." });    
     }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     return res.render("videos/edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
+    const { user: { _id } } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({_id:id});
     if(!video){
         return res.status(404).render("error/404", { pageTitle: "Video not found." });    
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -43,15 +54,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const { user: { _id } } = req.session;
     const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+
         return res.redirect("/");
     } catch(error){
         return res.status(400).render("videos/upload", { 
@@ -62,7 +80,15 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-   const { id } = req.params;
+    const { id } = req.params;
+    const { user: { _id } } = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("error/404", { pageTitle: "Video not found." });    
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
    await Video.findByIdAndDelete(id);
    return res.redirect("/");
 };
